+++ src/main/java/fr/tfl/store/model/ProfilModel.java
@@ -10,16 +10,15 @@ import fr.tfl.store.bean.Profil;
 
 public class ProfilModel implements Serializable {
 	
-	private Integer id;
+	private Long id;
 	private String profil;
 	private String label;
 	private int version;	
-	
-	
-	public Integer getId() {
+		
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getProfil() {
